// src/app/core/models/registro.model.ts
export interface Registro {
  id?: number;
  foj: string;
  nro: string;
  ano: string;
  bim?: string;
  nom1?: string;
  nom2?: string;
  cont?: string;
  materia?: string;
  fesc?: string;
  fechaCorregida?: string;
  req?: string;
  nBoleta?: string;
  aran?: string;
  fent?: string;
  abog?: string;
  rut1?: string;
  rut1d?: string;
  rut2?: string;
  rut2d?: string;
  letra?: string;
  matri?: string;
  matricero?: string;
  difCorr?: string;
  observacion?: string;
  createdAt?: string;
  updatedAt?: string;
}